requirements.txt
fastapi==0.95.1
python-ldap==3.4.3
uvicorn==0.22.0
gunicorn==20.1.0
python-dotenv==1.0.0